requirements.txt
python-dotenv
Flask
Groq
flask-cors
flasgger

